Checkpoint -- + Make PreferencePanel more Similar to ConfigPanel + Add Dummy Button for new feature
codev: Lee and Amanda
